feat: [website-header-formatting-004] - [Add focused regression coverage for the updated header formatting]
